pull out single-block behaviors which don't depend on gate behavior into new abstract class
