+++ src/main/java/name/turingcomplete/blocks/block/MEMORY_Cell_Block.java
@@ -54,6 +54,7 @@ public class MEMORY_Cell_Block extends AbstractSimpleGate {
 
     @Override
     protected void onNeighborUpdate(World world, BlockPos gatePos, BlockState gateState){
+        super.onNeighborUpdate(world, gatePos, gateState);
         world.setBlockState(gatePos, gateState.with(
             ENABLED,
             getInputActive(world, gatePos, gateState, getEnabledSide(gateState))));
